# openapi/teldrive.openapi.yaml
openapi: 3.0.0
info:
  title: TelDrive API
  version: 1.0.0
  summary: Telegram-backed cloud drive API
  description: >-
    Backend-first API for rclone and other TelDrive clients. Upload sessions are

    durable, API instances are stateless, and completed files are immediately
    readable.
...
paths:
...
  /v1/files/{fileId}/content:
    get:
      operationId: downloadFile
      description: Stream complete or partial file content.
      parameters:
        - $ref: "#/components/parameters/FileContentRequestHeaders.range"
        - $ref: "#/components/parameters/FileContentRequestHeaders.ifNoneMatch"
        - name: fileId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: The request has succeeded.
          headers:
            Content-Length:
              required: true
              schema:
                type: integer
                format: int64
            Accept-Ranges:
              required: true
              schema:
                type: string
                enum:
                  - bytes
            Etag:
              required: true
              schema:
                $ref: "#/components/schemas/ETag"
            Last-Modified:
              required: true
              schema:
                type: string
                format: http-date
            Content-Disposition:
              required: true
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
              x-ogen-raw-response: true
        "206":
          description: Successful
          headers:
            Content-Length:
              required: true
              schema:
                type: integer
                format: int64
            Content-Range:
              required: true
              schema:
                type: string
            Accept-Ranges:
              required: true
              schema:
                type: string
                enum:
                  - bytes
            Etag:
              required: true
              schema:
                $ref: "#/components/schemas/ETag"
            Last-Modified:
              required: true
              schema:
                type: string
                format: http-date
            Content-Disposition:
              required: true
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
              x-ogen-raw-response: true
        "304":
          description: The client has made a conditional request and the resource has not
            been modified.
        "401":
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "416":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      tags:
        - Files
      security:
        - BearerAuth: []
        - CookieAuth: []
        - ExternalApiKeyAuth: []
    head:
      operationId: headFile
      description: Return content metadata without streaming the file.
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: The request has succeeded.
          headers:
            Content-Length:
              required: true
              schema:
                type: integer
                format: int64
            Accept-Ranges:
              required: true
              schema:
                type: string
                enum:
                  - bytes
            Etag:
              required: true
              schema:
                $ref: "#/components/schemas/ETag"
            Last-Modified:
              required: true
              schema:
                type: string
                format: http-date
            Content-Disposition:
              required: true
              schema:
                type: string
        "401":
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      tags:
        - Files
      security:
        - BearerAuth: []
        - CookieAuth: []
        - ExternalApiKeyAuth: []
...
  /v1/public/shares/{token}/files/{fileId}/content:
    get:
      operationId: downloadPublicShareFile
      parameters:
        - $ref: "#/components/parameters/SharePasswordHeader"
        - $ref: "#/components/parameters/FileContentRequestHeaders.range"
        - $ref: "#/components/parameters/FileContentRequestHeaders.ifNoneMatch"
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: fileId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: The request has succeeded.
          headers:
            Content-Length:
              required: true
              schema:
                type: integer
                format: int64
            Accept-Ranges:
              required: true
              schema:
                type: string
                enum:
                  - bytes
            Etag:
              required: true
              schema:
                $ref: "#/components/schemas/ETag"
            Last-Modified:
              required: true
              schema:
                type: string
                format: http-date
            Content-Disposition:
              required: true
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
              x-ogen-raw-response: true
        "206":
          description: Successful
          headers:
            Content-Length:
              required: true
              schema:
                type: integer
                format: int64
            Content-Range:
              required: true
              schema:
                type: string
            Accept-Ranges:
              required: true
              schema:
                type: string
                enum:
                  - bytes
            Etag:
              required: true
              schema:
                $ref: "#/components/schemas/ETag"
            Last-Modified:
              required: true
              schema:
                type: string
                format: http-date
            Content-Disposition:
              required: true
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
              x-ogen-raw-response: true
        "304":
          description: The client has made a conditional request and the resource has not
            been modified.
        "401":
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "410":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "416":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      tags:
        - Public shares
      security:
        - {}
    head:
      operationId: headPublicShareFile
      parameters:
        - $ref: "#/components/parameters/SharePasswordHeader"
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: fileId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/Uuid"
      responses:
        "200":
          description: The request has succeeded.
          headers:
            Content-Length:
              required: true
              schema:
                type: integer
                format: int64
            Accept-Ranges:
              required: true
              schema:
                type: string
                enum:
                  - bytes
            Etag:
              required: true
              schema:
                $ref: "#/components/schemas/ETag"
            Last-Modified:
              required: true
              schema:
                type: string
                format: http-date
            Content-Disposition:
              required: true
              schema:
                type: string
        "401":
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "410":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      tags:
        - Public shares
      security:
        - {}
  /v1/public/shares/{token}/content:
    get:
      operationId: downloadPublicShare
      parameters:
        - $ref: "#/components/parameters/SharePasswordHeader"
        - $ref: "#/components/parameters/FileContentRequestHeaders.range"
        - $ref: "#/components/parameters/FileContentRequestHeaders.ifNoneMatch"
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The request has succeeded.
          headers:
            Content-Length:
              required: true
              schema:
                type: integer
                format: int64
            Accept-Ranges:
              required: true
              schema:
                type: string
                enum:
                  - bytes
            Etag:
              required: true
              schema:
                $ref: "#/components/schemas/ETag"
            Last-Modified:
              required: true
              schema:
                type: string
                format: http-date
            Content-Disposition:
              required: true
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
              x-ogen-raw-response: true
        "206":
          description: Successful
          headers:
            Content-Length:
              required: true
              schema:
                type: integer
                format: int64
            Content-Range:
              required: true
              schema:
                type: string
            Accept-Ranges:
              required: true
              schema:
                type: string
                enum:
                  - bytes
            Etag:
              required: true
              schema:
                $ref: "#/components/schemas/ETag"
            Last-Modified:
              required: true
              schema:
                type: string
                format: http-date
            Content-Disposition:
              required: true
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
              x-ogen-raw-response: true
        "304":
          description: The client has made a conditional request and the resource has not
            been modified.
        "401":
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "410":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "416":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      tags:
        - Public shares
      security:
        - {}
    head:
      operationId: headPublicShare
      parameters:
        - $ref: "#/components/parameters/SharePasswordHeader"
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The request has succeeded.
          headers:
            Content-Length:
              required: true
              schema:
                type: integer
                format: int64
            Accept-Ranges:
              required: true
              schema:
                type: string
                enum:
                  - bytes
            Etag:
              required: true
              schema:
                $ref: "#/components/schemas/ETag"
            Last-Modified:
              required: true
              schema:
                type: string
                format: http-date
            Content-Disposition:
              required: true
              schema:
                type: string
        "401":
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "410":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          description: Client error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Service unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
      tags:
        - Public shares
      security:
        - {}
...
components:
  parameters:
...
    FileContentRequestHeaders.range:
      name: Range
      in: header
      required: false
      schema:
        type: string
    FileContentRequestHeaders.ifNoneMatch:
      name: If-None-Match
      in: header
      required: false
      schema:
        $ref: "#/components/schemas/ETag"
...
    SharePasswordHeader:
      name: X-Share-Password
      in: header
      required: false
      schema:
        type: string
...
  schemas:
...
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-readable error message.
        details:
          type: object
          additionalProperties: {}
          description: Optional structured context for the caller.
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    Uuid:
      type: string
      format: uuid
...
    ETag:
      type: string
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: teldrive_access
    BearerAuth:
      type: http
      scheme: Bearer
    ExternalApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    EventTicketAuth:
      type: apiKey
      in: query
      name: ticket
servers:
  - url: /api
    description: Current TelDrive server
    variables: {}